Improved code according to @petertodd insights
